nav bar temp fix
I could not get the drop-down menus to work in the navbar.  After much
testing, I determined that they have never worked -- the CSS needs to
be fixed.  (I tried many adjustments to the HTML and then finally
reinstalled Luke Hollis's nav bar, which had a drop-down menu in the
HTML.  It didn't work; we must have missed it in the testing.) So
instead of having a broken nav bar that looks like the working one at
copticscriptorium.org, I pared it down to have only HOME and ABOUT.  No
dropdowns.